put festival_helpers in scripts/ as well, and adjust sys.path accordingly. maybe rename it too? Fixes #38
