+++ docs/alkmaarse-eigenste/2026/about.html
@@ -124,7 +124,7 @@
                 <p class="festival-description" style="font-size: 0.95em; opacity: 0.85; margin-top: 0.5rem; max-width: 800px;">A single-day discovery festival in Alkmaar showcasing local and emerging artists.</p>
                 
                 <p class="subtitle" style="font-size: 0.8em; opacity: 0.7; margin-top: 0.5rem;">
-                    About page generated: December 14, 2025 20:17 UTC
+                    About page generated: December 14, 2025 20:27 UTC
                      | <a href="https://www.podiumvictorie.nl/programma/alkmaarseigenste2026/" target="_blank" rel="noopener noreferrer" style="color: inherit; text-decoration: none;">Festival Site</a>
                 </p>
             </div>
@@ -133,10 +133,10 @@
         <div class="container-fluid" style="max-width: 1400px; padding: 2rem 1rem;">
             <section class="mb-5">
                 <h2 style="margin-bottom: 1.5rem; font-size: 1.75rem; font-weight: 600;">Festival Profile</h2>
-                <div class="profile-text"><p><strong>Alkmaarse Eigenste 2026: A Celebration of Local Talent and Eclectic Sounds</strong></p>
-<p>The Alkmaarse Eigenste 2026 festival brings together 25 exceptional artists, all hailing from the Netherlands, to showcase the rich and diverse music scene of the region. With a focus on alternative and indie genres, the lineup features a vibrant mix of sounds, including alternative (6 artists), rock (4), and alternative pop (3). The festival takes pride in its eclectic selection, spanning everything from dream pop and indie folk to blues rock and reggae, ensuring a unique experience for every music lover.</p>
-<p>Diversity is a key theme this year, with a balance of male (9), female (7), and band-led acts (4), as well as a range of musical styles that reflect the creativity and individuality of the artists. While the majority of performers are not people of color (19), the inclusion of a POC artist and the variety of genres, from jazz and world music to punk and alt-pop-noir, highlights the festival’s commitment to representing different voices and perspectives within its lineup.</p>
-<p>With an impressive average rating of 10.0 for its carefully curated roster and all artists featured on Spotify, Alkmaarse Eigenste 2026 is set to deliver an unforgettable celebration of artistry and innovation. This festival stands as a vibrant platform for exploring the diverse sounds of the Netherlands, uniting music lovers under the banner of creativity and local pride.</p></div>
+                <div class="profile-text"><p><strong>Alkmaarse Eigenste 2026: A Celebration of Dutch Talent and Musical Diversity</strong></p>
+<p>Alkmaarse Eigenste 2026 brings together an eclectic mix of 25 artists, all proudly hailing from the Netherlands. Showcasing the country's rich musical landscape, the festival spans a wide array of genres, from alternative and rock to indie pop, soul, and even unique blends like swampy roots, alt-pop-noir, and synthesizer rock. With alternative music leading the charge, the lineup reflects a strong focus on creativity and innovation, offering something for every music lover.</p>
+<p>The festival emphasizes inclusivity and diversity, with a near-equal representation of male (9) and female (7) artists, alongside several bands and acts with undisclosed gender identities. While the majority of performers are not people of color, Alkmaarse Eigenste is taking small steps toward broader representation, highlighting one artist of color among the lineup. Every artist featured this year has a presence on Spotify, ensuring fans can engage with their music before and after the festival.</p>
+<p>With an impressive average rating of 10.0 based on early reviews, Alkmaarse Eigenste 2026 promises an unforgettable experience for attendees. The festival blends local pride with an appreciation for artistic diversity, creating a platform that celebrates both established and emerging talents across a spectrum of genres, styles, and influences.</p></div>
             </section>
             
             <section>
