+++ docs/alkmaarse-eigenste/2026/about.json
@@ -2,7 +2,7 @@
   "festival": "alkmaarse-eigenste",
   "name": "Alkmaarse Eigenste",
   "year": 2026,
-  "generated_at": "2025-12-14T20:17:41.821961+00:00",
+  "generated_at": "2025-12-14T20:27:18.314714+00:00",
   "config_properties": {
     "name": "Alkmaarse Eigenste",
     "year": 2026,
@@ -91,5 +91,5 @@
     "year": 2026
   },
   "previous_year_comparison": {},
-  "ai_profile": "**Alkmaarse Eigenste 2026: A Celebration of Local Talent and Eclectic Sounds**\n\nThe Alkmaarse Eigenste 2026 festival brings together 25 exceptional artists, all hailing from the Netherlands, to showcase the rich and diverse music scene of the region. With a focus on alternative and indie genres, the lineup features a vibrant mix of sounds, including alternative (6 artists), rock (4), and alternative pop (3). The festival takes pride in its eclectic selection, spanning everything from dream pop and indie folk to blues rock and reggae, ensuring a unique experience for every music lover.\n\nDiversity is a key theme this year, with a balance of male (9), female (7), and band-led acts (4), as well as a range of musical styles that reflect the creativity and individuality of the artists. While the majority of performers are not people of color (19), the inclusion of a POC artist and the variety of genres, from jazz and world music to punk and alt-pop-noir, highlights the festival’s commitment to representing different voices and perspectives within its lineup.\n\nWith an impressive average rating of 10.0 for its carefully curated roster and all artists featured on Spotify, Alkmaarse Eigenste 2026 is set to deliver an unforgettable celebration of artistry and innovation. This festival stands as a vibrant platform for exploring the diverse sounds of the Netherlands, uniting music lovers under the banner of creativity and local pride."
+  "ai_profile": "**Alkmaarse Eigenste 2026: A Celebration of Dutch Talent and Musical Diversity**  \n\nAlkmaarse Eigenste 2026 brings together an eclectic mix of 25 artists, all proudly hailing from the Netherlands. Showcasing the country's rich musical landscape, the festival spans a wide array of genres, from alternative and rock to indie pop, soul, and even unique blends like swampy roots, alt-pop-noir, and synthesizer rock. With alternative music leading the charge, the lineup reflects a strong focus on creativity and innovation, offering something for every music lover.  \n\nThe festival emphasizes inclusivity and diversity, with a near-equal representation of male (9) and female (7) artists, alongside several bands and acts with undisclosed gender identities. While the majority of performers are not people of color, Alkmaarse Eigenste is taking small steps toward broader representation, highlighting one artist of color among the lineup. Every artist featured this year has a presence on Spotify, ensuring fans can engage with their music before and after the festival.  \n\nWith an impressive average rating of 10.0 based on early reviews, Alkmaarse Eigenste 2026 promises an unforgettable experience for attendees. The festival blends local pride with an appreciation for artistic diversity, creating a platform that celebrates both established and emerging talents across a spectrum of genres, styles, and influences."
 }
