+++ scripts/fetch_festival_data.py
@@ -16,13 +16,14 @@ import sys
 from pathlib import Path
 
 # Add parent directory to sys.path to import festival_helpers
-sys.path.insert(0, str(Path(__file__).parent.parent))
+import sys
+sys.path.insert(0, str(Path(__file__).parent))
 
 import csv
 import json
 import argparse
 from typing import Dict, List
-from festival_helpers import (
+from helpers import (
     FestivalScraper,
     get_festival_config,
     artist_name_to_slug,
