+++ scripts/helpers/ai_client.py
