+++ scripts/helpers/generate_charts.py
@@ -6,9 +6,9 @@ import csv
 from pathlib import Path
 import json
 import sys
-sys.path.append(str(Path(__file__).parent.parent.parent))
+# helpers module is in the same directory
 
-from festival_helpers.menu import generate_hamburger_menu
+from helpers.menu import generate_hamburger_menu
 
 def load_festival_data(csv_path):
     """Load and analyze festival data from CSV."""
