+++ scripts/helpers/text_utils.py
