+++ scripts/update_lineup.py
@@ -10,11 +10,12 @@ import sys
 from pathlib import Path
 
 # Add parent directory to sys.path to import festival_helpers
-sys.path.insert(0, str(Path(__file__).parent.parent))
+import sys
+sys.path.insert(0, str(Path(__file__).parent))
 
 import re
 from festival_tracker import FestivalTracker
-from festival_helpers import FestivalScraper, get_festival_config
+from helpers import FestivalScraper, get_festival_config
 from typing import List, Dict
