+++ tests/test_scraper.py
@@ -3,7 +3,10 @@ Tests for festival_helpers/scraper.py
 """
 import pytest
 from unittest.mock import Mock, patch, MagicMock
-from festival_helpers.scraper import FestivalScraper
+import sys
+from pathlib import Path
+sys.path.insert(0, str(Path(__file__).parent.parent / "scripts"))
+from helpers.scraper import FestivalScraper
 
 
 class TestFestivalScraper:
